refactor(i18n): adopt folder structure prior to build/publish setup
