+++ src/common/i18n/src/consts.ts
