+++ src/common/i18n/src/polyglot/LocaleAwarePolyglot.test.ts
@@ -6,7 +6,7 @@ import {
   SUPPORTED_LOCALES,
   SupportedLocale,
 } from '@i18n/consts';
-import { LocaleAwarePolyglot } from '@i18n/LocaleAwarePolyglot';
+import { LocaleAwarePolyglot } from '@i18n/polyglot/LocaleAwarePolyglot';
 import Polyglot from 'node-polyglot';
 import {
   beforeEach,
